langchainXinetum/pyproject.toml
[project]
name = "ia-langchain-api-4-genai-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "build>=1.2.2.post1",
    "langchain>=0.3.23",
    "langchain-core>=0.3.51",
    "python-dotenv>=1.1.0",
]
